Make `main` breakpoint temporary
This breakpoint does not appear in the UI, so it is confusing when the code stops here multiple times (e.g., when `main` is a loop like in this project). Adding and removing a UI breakpoint does not remove the actual breakpoint. This change makes it a temporary breakpoint, which is removed when first visited. This also allows users to place one more breakpoint they otherwise couldn't use (limit is ~6 or so).